Move project classification action into classify controller
More crud, switching to separate controller for the create action add
unnecessary complexity.
